added hacky 5.2+ workaround example to readme
IDK if this would be the obvious way for others to do it, but I used to have a main and a "true main" file in Love2d projects because having to wrap my main module in one big function was just ugly. I think this is a nice way to abuse the debug library.